+++ Proyecto_App_Bancaria/nbproject/project.properties
@@ -74,7 +74,7 @@ jlink.additionalmodules=
 jlink.additionalparam=
 jlink.launcher=true
 jlink.launcher.name=Proyecto_App_Bancaria
-main.class=gui.Inicio
+main.class=guiUsuarios.Inicio
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
