+++ Proyecto_App_Bancaria/src/guiAdministradores/SolicitudesDesbloqueo.java
@@ -1,11 +1,13 @@
-package gui;
+package guiAdministradores;
 
+import guiAdministradores.PrincipalAdmins;
 import Personas.Administradores;
 import Personas.Usuarios;
 import Sedes.SedeCentral;
 import Sedes.SedeCiudadColon;
 import Sedes.SedePuriscal;
 import Sedes.SedeSanPedro;
+import guiUsuarios.Inicio;
 import java.awt.Color;
 import java.awt.event.KeyEvent;
 import javax.swing.JOptionPane;
@@ -79,6 +81,8 @@ public class SolicitudesDesbloqueo extends javax.swing.JFrame {
         itemSolicitudesDesbloqueo = new javax.swing.JMenuItem();
         jSeparator2 = new javax.swing.JPopupMenu.Separator();
         itemBloquearCuenta = new javax.swing.JMenuItem();
+        jSeparator5 = new javax.swing.JPopupMenu.Separator();
+        itemDesbloquearCuenta = new javax.swing.JMenuItem();
         menuGestionFinanciera = new javax.swing.JMenu();
         menuGestionSedes = new javax.swing.JMenu();
         itemSedePuriscal = new javax.swing.JMenuItem();
@@ -241,6 +245,16 @@ public class SolicitudesDesbloqueo extends javax.swing.JFrame {
             }
         });
         menuBloqueoDesbloqueo.add(itemBloquearCuenta);
+        menuBloqueoDesbloqueo.add(jSeparator5);
+
+        itemDesbloquearCuenta.setFont(new java.awt.Font("Segoe UI", 0, 14)); // NOI18N
+        itemDesbloquearCuenta.setText("Desbloquear cuenta");
+        itemDesbloquearCuenta.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                itemDesbloquearCuentaActionPerformed(evt);
+            }
+        });
+        menuBloqueoDesbloqueo.add(itemDesbloquearCuenta);
 
         menuGestionUsuarios.add(menuBloqueoDesbloqueo);
 
@@ -410,6 +424,12 @@ public class SolicitudesDesbloqueo extends javax.swing.JFrame {
         }
     }//GEN-LAST:event_txtClaveNumKeyTyped
 
+    private void itemDesbloquearCuentaActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_itemDesbloquearCuentaActionPerformed
+        DesbloquearCuenta desbloquear = new DesbloquearCuenta(this.cedula);
+        desbloquear.setVisible(true);
+        this.dispose();
+    }//GEN-LAST:event_itemDesbloquearCuentaActionPerformed
+
     /**
      * @param args the command line arguments
      */
@@ -450,6 +470,7 @@ public class SolicitudesDesbloqueo extends javax.swing.JFrame {
     private javax.swing.JMenuBar barraMenu;
     private javax.swing.JMenuItem itemBienvenida;
     private javax.swing.JMenuItem itemBloquearCuenta;
+    private javax.swing.JMenuItem itemDesbloquearCuenta;
     private javax.swing.JMenuItem itemIrInicio;
     private javax.swing.JMenuItem itemSedeCiudadColon;
     private javax.swing.JMenuItem itemSedePuriscal;
@@ -464,6 +485,7 @@ public class SolicitudesDesbloqueo extends javax.swing.JFrame {
     private javax.swing.JPopupMenu.Separator jSeparator2;
     private javax.swing.JPopupMenu.Separator jSeparator3;
     private javax.swing.JPopupMenu.Separator jSeparator4;
+    private javax.swing.JPopupMenu.Separator jSeparator5;
     private javax.swing.JLabel jblDesbloquear;
     private javax.swing.JLabel jblSoloNumeros;
     private javax.swing.JMenu menuBloqueoDesbloqueo;
