+++ Proyecto_App_Bancaria/src/guiUsuarios/ConfigurarPerfil.java
@@ -1,4 +1,4 @@
-package gui;
+package guiUsuarios;
 
 import Personas.Usuarios;
 import RegistroDatos.DatosRegistrados;
