+++ Proyecto_App_Bancaria/src/guiUsuarios/Inicio.java
@@ -1,5 +1,5 @@
 
-package gui;
+package guiUsuarios;
 
 import java.awt.Color;
