+++ Proyecto_App_Bancaria/src/guiUsuarios/Registro.java
@@ -1,4 +1,4 @@
-package gui;
+package guiUsuarios;
 
 import Personas.Administradores;
 import Personas.Usuarios;
